feat(mui): add HambergerMenu button and fix __Internal__HasHeaderWithThemedContext
